修正了几个今天提交后的警告信息

+++ 20-ResizeWindowEmpty/20-ResizeWindowEmpty.cpp
@@ -671,8 +671,8 @@ void OnSize(UINT width, UINT height, bool minimized)
         // 8、更新视口大小和齐次包围盒大小
         g_stGPUStatus.m_stViewPort.TopLeftX = 0.0f;
         g_stGPUStatus.m_stViewPort.TopLeftY = 0.0f;
-        g_stGPUStatus.m_stViewPort.Width = g_stGPUStatus.m_iWndWidth;
-        g_stGPUStatus.m_stViewPort.Height = g_stGPUStatus.m_iWndHeight;
+        g_stGPUStatus.m_stViewPort.Width = (float)g_stGPUStatus.m_iWndWidth;
+        g_stGPUStatus.m_stViewPort.Height = (float)g_stGPUStatus.m_iWndHeight;
 
         g_stGPUStatus.m_stScissorRect.left = static_cast<LONG>(g_stGPUStatus.m_stViewPort.TopLeftX);
         g_stGPUStatus.m_stScissorRect.right = static_cast<LONG>(g_stGPUStatus.m_stViewPort.TopLeftX + g_stGPUStatus.m_stViewPort.Width);
